+++ controllers/ordersController.js
@@ -70,6 +70,13 @@ const updating = async (req, res, next) => {
 const deleting = async (req, res, next) => {
   try {
     const id = req.params.id;
+
+    // check OrderId in table OrderItem
+    const orderIdin = await OrderId.findOne({ where: { orderId: id } });
+    if (orderIdin) {
+      return res.status(400).json({ message: "Cannot delete Order because it is associated with OrderItem. First you need to delete OrderId in OrderItem" });
+    }    
+    
     const deleted = await Order.destroy({ where: { id } });
     if (deleted === 0) {
       return next(
